+++ src/Domain/DataMapper/TeamMapper.php
@@ -5,6 +5,7 @@
  */
 namespace Clanify\Domain\DataMapper;
 
+use Clanify\Domain\Entity\IEntity;
 use Clanify\Domain\Entity\Team;
 
 /**
@@ -19,20 +20,29 @@ use Clanify\Domain\Entity\Team;
 class TeamMapper extends DataMapper
 {
     /**
-     * The name of the table which will be used.
+     * TeamMapper constructor.
+     * @param \PDO $pdo The PDO object.
      * @since 0.0.1-dev
-     * @var string
      */
-    private $table = 'team';
+    public function __construct(\PDO $pdo)
+    {
+        $this->table = 'team';
+        parent::__construct($pdo);
+    }
 
     /**
      * Method to create a Team on database.
-     * @param Team $team The Entity of the Team.
-     * @return bool The state if the Team was successfully created.
+     * @param IEntity $team The Team Entity.
+     * @return bool The state if the Team Entity was successfully created.
      * @since 0.0.1-dev
      */
-    private function create(Team $team)
+    public function create(IEntity $team)
     {
+        //check if a Team is available.
+        if (!($team instanceof Team)) {
+            return false;
+        }
+
         //create and set the sql query.
         $sql = 'INSERT INTO '.$this->table.' (name, tag, website) VALUES (:name, :tag, :website);';
         $sth = $this->pdo->prepare($sql);
@@ -48,12 +58,17 @@ class TeamMapper extends DataMapper
 
     /**
      * Method to delete a Team on database.
-     * @param Team $team The Entity of the Team.
-     * @return bool The state if the Team was successfully deleted.
+     * @param IEntity $team The Team Entity.
+     * @return bool The state if the Team Entity was successfully deleted.
      * @since 0.0.1-dev
      */
-    public function delete(Team $team)
+    public function delete(IEntity $team)
     {
+        //check if a Team is available.
+        if (!($team instanceof Team)) {
+            return false;
+        }
+
         //create and set the sql query.
         $sql = 'DELETE FROM '.$this->table.' WHERE id = :id;';
         $sth = $this->pdo->prepare($sql);
@@ -66,24 +81,46 @@ class TeamMapper extends DataMapper
     }
 
     /**
-     * Method to save a Team on database.
-     * @param Team $team The Entity of the Team.
-     * @return bool The state if the Team was successfully saved.
+     * Method to find Team Entities by a SQL condition.
+     * @param string $condition The SQL condition to filter the Team Entities.
+     * @return array An array with all found Team Entities.
      * @since 0.0.1-dev
      */
-    public function save(Team $team)
+    public function find($condition = '')
     {
+        return $this->findForEntity($condition, new Team());
+    }
+
+    /**
+     * Method to save a Team on database.
+     * @param IEntity $team The Team Entity.
+     * @return bool The state if the Team Entity was successfully saved.
+     * @since 0.0.1-dev
+     */
+    public function save(IEntity $team)
+    {
+        //check if a Team is available.
+        if (!($team instanceof Team)) {
+            return false;
+        }
+
+        //create or update the Team.
         return ($team->id > 0) ? $this->update($team) : $this->create($team);
     }
 
     /**
      * Method to update a Team on database.
-     * @param Team $team The Entity of the Team.
-     * @return bool The state if the Team was successfully updated.
+     * @param IEntity $team The Team Entity.
+     * @return bool The state if the Team Entity was successfully updated.
      * @since 0.0.1-dev
      */
-    private function update(Team $team)
+    public function update(IEntity $team)
     {
+        //check if a Team is available.
+        if (!($team instanceof Team)) {
+            return false;
+        }
+
         //create and set the sql query.
         $sql = 'UPDATE '.$this->table.' SET name = :name, tag = :tag, website = :website WHERE id = :id;';
         $sth = $this->pdo->prepare($sql);
